refactor into redact.go and fix unit test
